Add coefficient type typealias to BLIT for testing
And add a [[main]] graph (commented out), to run
the oscillator by itself.

The addition of typealias should be dropped. It's
committed here to facilitate comparision and trouble shooting.

The problem? It blows up when type is float64. I don't know
what that would be, but I don't think it's a problem with
my code.